fix: moved providers interface to model package, provider selection logic to provider package, and minor fixes
