+++ go.mod
@@ -11,5 +11,5 @@ require (
 require (
 	github.com/samber/lo v1.38.1 // indirect
 	golang.org/x/exp v0.0.0-20230522175609-2e198f4a06a1 // indirect
-	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.1
 )
